flask-app/requirements.txt
flask
flask_bootstrap
flask_login
flask_wtf
argon2-cffi
mysql-connector-python